DRC-2559: Allow variable clients to specify data frequency. does not fixbamboo

+++ RobotDataCommunication/src/us/ihmc/robotDataCommunication/logger/YoVariableLoggerListener.java
@@ -319,7 +319,7 @@ public class YoVariableLoggerListener implements YoVariablesUpdatedListener
    }
 
    @Override
-   public long getDisplayOneInNPackets()
+   public int getDisplayOneInNPackets()
    {
       return 1;
    }
